requirements.txt
# Python dependencies for API testing
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
allure-pytest==2.13.2
requests==2.31.0
python-dotenv==1.0.0
jsonschema==4.20.0
pydantic==2.5.3
